refactor: move global FF lowering into frontend
