--- crates/celox/src/parser/ff.rs
@@ -1,1 +1,0 @@
-pub use celox_frontend_veryl::ff::*;
